binance-scalping-bot/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
ccxt==4.2.97
binance-futures-connector==4.0.0
websockets==12.0
python-dotenv==1.0.1
httpx==0.28.1
pymysql==1.1.1
